backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gitpython
sentence-transformers
qdrant-client
openai
requests
google-generativeai
